Different functions that perform binning.

Different functions that perform binning.

cut_at(breaks)

nearest(centers)

bin_by_ntile(nbins)

bin_by_eqcut(nbins)

bin_by_pam(nbins)

bin_by_classInt(style, nbins = NULL)

Arguments

breaks

A numeric vector of values that designate cut points between bins.

centers

A numeric vector of values that designate the center of each bin.

nbins

The number of bins to split the data into.

style

a binning style (see classIntervals for details).

Value

Each of these functions returns a function of a single numeric vector `x` that assigns each value of `x` to a bin.

Examples

x <- c(rnorm(10, 1, 1), rnorm(10, 3, 2), rnorm(20, 5, 3))
centers <- c(1, 3, 5)
nearest(centers)(x)
#>  [1] 1 1 1 3 1 1 1 1 1 1 5 3 3 1 3 3 1 3 5 5 3 5 3 5 3 5 5 5 5 5 1 5 3 5 5 5 1 3
#> [39] 3 5

breaks <- c(2, 4)
cut_at(breaks)(x)
#>  [1] "[0.0656,2)" "[0.0656,2)" "[0.0656,2)" "[2,4)"      "[0.0656,2)"
#>  [6] "[0.0656,2)" "[0.0656,2)" "[0.0656,2)" "[0.0656,2)" "[0.0656,2)"
#> [11] "[4,11.3]"   "[2,4)"      "[2,4)"      "[0.0656,2)" "[2,4)"     
#> [16] "[2,4)"      "[0.0656,2)" "[2,4)"      "[4,11.3]"   "[4,11.3]"  
#> [21] "[2,4)"      "[4,11.3]"   "[2,4)"      "[4,11.3]"   "[2,4)"     
#> [26] "[4,11.3]"   "[4,11.3]"   "[4,11.3]"   "[4,11.3]"   "[4,11.3]"  
#> [31] "[0.0656,2)" "[4,11.3]"   "[2,4)"      "[4,11.3]"   "[4,11.3]"  
#> [36] "[4,11.3]"   "[0.0656,2)" "[2,4)"      "[2,4)"      "[4,11.3]"  

bin_by_eqcut(nbins=4)(x)
#>  [1] "[0.0656,1.33)" "[0.0656,1.33)" "[0.0656,1.33)" "[1.33,3.2)"   
#>  [5] "[0.0656,1.33)" "[0.0656,1.33)" "[0.0656,1.33)" "[0.0656,1.33)"
#>  [9] "[0.0656,1.33)" "[1.33,3.2)"    "[6.32,11.3]"   "[1.33,3.2)"   
#> [13] "[1.33,3.2)"    "[1.33,3.2)"    "[3.2,6.32)"    "[3.2,6.32)"   
#> [17] "[1.33,3.2)"    "[1.33,3.2)"    "[6.32,11.3]"   "[3.2,6.32)"   
#> [21] "[1.33,3.2)"    "[6.32,11.3]"   "[3.2,6.32)"    "[6.32,11.3]"  
#> [25] "[3.2,6.32)"    "[3.2,6.32)"    "[6.32,11.3]"   "[6.32,11.3]"  
#> [29] "[6.32,11.3]"   "[3.2,6.32)"    "[0.0656,1.33)" "[6.32,11.3]"  
#> [33] "[1.33,3.2)"    "[6.32,11.3]"   "[3.2,6.32)"    "[3.2,6.32)"   
#> [37] "[0.0656,1.33)" "[3.2,6.32)"    "[1.33,3.2)"    "[6.32,11.3]"  
bin_by_ntile(nbins=4)(x)
#>  [1] 1 1 1 2 1 1 1 1 1 2 4 2 2 2 3 3 2 2 4 3 2 4 3 4 3 3 4 4 4 3 1 4 2 4 3 3 1 3
#> [39] 2 4

# \donttest{
bin_by_pam(nbins=4)(x)
#>  [1]  0.6055946  0.6055946  0.6055946  3.2291396  0.6055946  0.6055946
#>  [7]  0.6055946  0.6055946  0.6055946  0.6055946  6.9565846  3.2291396
#> [13]  3.2291396  0.6055946  3.2291396  3.2291396  0.6055946  3.2291396
#> [19]  6.9565846  3.2291396  3.2291396  6.9565846  3.2291396  6.9565846
#> [25]  3.2291396  3.2291396  6.9565846 10.3311046  6.9565846  6.9565846
#> [31]  0.6055946 10.3311046  3.2291396  6.9565846  3.2291396  6.9565846
#> [37]  0.6055946  3.2291396  3.2291396  6.9565846
bin_by_classInt("pretty", nbins=4)(x)
#>  [1] "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"  
#>  [8] "[0,5)"   "[0,5)"   "[0,5)"   "[5,10)"  "[0,5)"   "[0,5)"   "[0,5)"  
#> [15] "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[5,10)"  "[0,5)"   "[0,5)"  
#> [22] "[5,10)"  "[0,5)"   "[5,10)"  "[0,5)"   "[0,5)"   "[5,10)"  "[10,15]"
#> [29] "[5,10)"  "[5,10)"  "[0,5)"   "[10,15]" "[0,5)"   "[5,10)"  "[0,5)"  
#> [36] "[5,10)"  "[0,5)"   "[0,5)"   "[0,5)"   "[5,10)" 
# }