Different functions that perform binning.

Different functions that perform binning.

cut_at(breaks)

nearest(centers)

bin_by_ntile(nbins)

bin_by_eqcut(nbins)

bin_by_pam(nbins)

bin_by_classInt(style, nbins = NULL)

Arguments

breaks
A numeric vector of values that designate cut points between bins.

centers
A numeric vector of values that designate the center of each bin.

nbins
The number of bins to split the data into.

style
a binning style (see classIntervals for details).

Value

Each of these functions returns a function of a single numeric vector `x` that assigns each value of `x` to a bin.

Examples

x <- c(rnorm(10, 1, 1), rnorm(10, 3, 2), rnorm(20, 5, 3))
centers <- c(1, 3, 5)
nearest(centers)(x)

#>  [1] 1 1 1 1 1 3 1 1 1 1 1 5 5 1 5 1 1 3 5 1 5 5 1 5 5 5 3 5 3 1 5 5 5 5 5 5 1 5
#> [39] 5 5

breaks <- c(2, 4)
cut_at(breaks)(x)

#>  [1] "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[2,4)"    
#>  [7] "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[-1.44,2)" "[4,10.7]" 
#> [13] "[4,10.7]"  "[-1.44,2)" "[4,10.7]"  "[-1.44,2)" "[-1.44,2)" "[2,4)"    
#> [19] "[4,10.7]"  "[-1.44,2)" "[4,10.7]"  "[4,10.7]"  "[-1.44,2)" "[4,10.7]" 
#> [25] "[4,10.7]"  "[4,10.7]"  "[2,4)"     "[4,10.7]"  "[2,4)"     "[-1.44,2)"
#> [31] "[4,10.7]"  "[4,10.7]"  "[4,10.7]"  "[4,10.7]"  "[4,10.7]"  "[4,10.7]" 
#> [37] "[-1.44,2)" "[4,10.7]"  "[4,10.7]"  "[4,10.7]" 

bin_by_eqcut(nbins=4)(x)

#>  [1] "[-1.44,0.935)" "[0.935,2.71)"  "[-1.44,0.935)" "[0.935,2.71)" 
#>  [5] "[0.935,2.71)"  "[0.935,2.71)"  "[-1.44,0.935)" "[-1.44,0.935)"
#>  [9] "[-1.44,0.935)" "[-1.44,0.935)" "[0.935,2.71)"  "[2.71,5.38)"  
#> [13] "[5.38,10.7]"   "[-1.44,0.935)" "[2.71,5.38)"   "[-1.44,0.935)"
#> [17] "[0.935,2.71)"  "[2.71,5.38)"   "[2.71,5.38)"   "[0.935,2.71)" 
#> [21] "[5.38,10.7]"   "[5.38,10.7]"   "[0.935,2.71)"  "[5.38,10.7]"  
#> [25] "[5.38,10.7]"   "[2.71,5.38)"   "[0.935,2.71)"  "[2.71,5.38)"  
#> [29] "[0.935,2.71)"  "[-1.44,0.935)" "[5.38,10.7]"   "[5.38,10.7]"  
#> [33] "[5.38,10.7]"   "[5.38,10.7]"   "[2.71,5.38)"   "[2.71,5.38)"  
#> [37] "[-1.44,0.935)" "[2.71,5.38)"   "[2.71,5.38)"   "[5.38,10.7]"  

bin_by_ntile(nbins=4)(x)

#>  [1] 1 2 1 2 2 2 1 1 1 1 2 3 4 1 3 1 2 3 3 2 4 4 2 4 4 3 2 3 2 1 4 4 4 4 3 3 1 3
#> [39] 3 4

# \donttest{
bin_by_pam(nbins=4)(x)

#>  [1] -0.7302624  1.2553171 -0.7302624  1.2553171  1.2553171  1.2553171
#>  [7] -0.7302624  1.2553171  1.2553171  1.2553171  1.2553171  4.5980090
#> [13]  7.8060896 -0.7302624  4.5980090 -0.7302624  1.2553171  1.2553171
#> [19]  4.5980090  1.2553171  7.8060896  4.5980090  1.2553171  7.8060896
#> [25]  7.8060896  4.5980090  1.2553171  4.5980090  1.2553171  1.2553171
#> [31]  7.8060896  4.5980090  4.5980090  7.8060896  4.5980090  4.5980090
#> [37] -0.7302624  4.5980090  4.5980090  7.8060896

bin_by_classInt("pretty", nbins=4)(x)

#>  [1] "[-5,0)"  "[0,5)"   "[-5,0)"  "[0,5)"   "[0,5)"   "[0,5)"   "[-5,0)" 
#>  [8] "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[5,10)"  "[-5,0)" 
#> [15] "[0,5)"   "[-5,0)"  "[0,5)"   "[0,5)"   "[0,5)"   "[0,5)"   "[5,10)" 
#> [22] "[5,10)"  "[0,5)"   "[5,10)"  "[10,15]" "[0,5)"   "[0,5)"   "[0,5)"  
#> [29] "[0,5)"   "[0,5)"   "[5,10)"  "[5,10)"  "[5,10)"  "[5,10)"  "[5,10)" 
#> [36] "[0,5)"   "[-5,0)"  "[0,5)"   "[0,5)"   "[5,10)" 

# }