Different functions that perform binning.

Different functions that perform binning.

cut_at(breaks)

nearest(centers)

bin_by_ntile(nbins)

bin_by_eqcut(nbins)

bin_by_pam(nbins)

bin_by_classInt(style, nbins = NULL)

Arguments

breaks

A numeric vector of values that designate cut points between bins.

centers

A numeric vector of values that designate the center of each bin.

nbins

The number of bins to split the data into.

style

a binning style (see classIntervals for details).

Value

Each of these functions returns a function of a single numeric vector `x` that assigns each value of `x` to a bin.

Examples

x <- c(rnorm(10, 1, 1), rnorm(10, 3, 2), rnorm(20, 5, 3))
centers <- c(1, 3, 5)
nearest(centers)(x)
#>  [1] 1 1 1 3 1 1 1 1 1 1 1 5 3 1 1 5 5 3 5 1 5 5 5 5 3 5 1 5 5 5 5 5 3 5 5 5 5 1
#> [39] 5 5

breaks <- c(2, 4)
cut_at(breaks)(x)
#>  [1] "[-1.29,2)" "[-1.29,2)" "[-1.29,2)" "[2,4)"     "[-1.29,2)" "[-1.29,2)"
#>  [7] "[-1.29,2)" "[-1.29,2)" "[-1.29,2)" "[-1.29,2)" "[-1.29,2)" "[4,9.9]"  
#> [13] "[2,4)"     "[-1.29,2)" "[-1.29,2)" "[4,9.9]"   "[4,9.9]"   "[2,4)"    
#> [19] "[4,9.9]"   "[-1.29,2)" "[4,9.9]"   "[4,9.9]"   "[4,9.9]"   "[4,9.9]"  
#> [25] "[2,4)"     "[4,9.9]"   "[-1.29,2)" "[4,9.9]"   "[4,9.9]"   "[4,9.9]"  
#> [31] "[4,9.9]"   "[4,9.9]"   "[2,4)"     "[4,9.9]"   "[4,9.9]"   "[4,9.9]"  
#> [37] "[4,9.9]"   "[-1.29,2)" "[4,9.9]"   "[4,9.9]"  

bin_by_eqcut(nbins=4)(x)
#>  [1] "[-1.29,0.814)" "[-1.29,0.814)" "[-1.29,0.814)" "[0.814,3.85)" 
#>  [5] "[0.814,3.85)"  "[0.814,3.85)"  "[0.814,3.85)"  "[-1.29,0.814)"
#>  [9] "[-1.29,0.814)" "[0.814,3.85)"  "[0.814,3.85)"  "[3.85,6.68)"  
#> [13] "[0.814,3.85)"  "[-1.29,0.814)" "[-1.29,0.814)" "[3.85,6.68)"  
#> [17] "[3.85,6.68)"   "[0.814,3.85)"  "[3.85,6.68)"   "[-1.29,0.814)"
#> [21] "[3.85,6.68)"   "[6.68,9.9]"    "[3.85,6.68)"   "[6.68,9.9]"   
#> [25] "[0.814,3.85)"  "[3.85,6.68)"   "[-1.29,0.814)" "[3.85,6.68)"  
#> [29] "[3.85,6.68)"   "[6.68,9.9]"    "[6.68,9.9]"    "[6.68,9.9]"   
#> [33] "[0.814,3.85)"  "[3.85,6.68)"   "[6.68,9.9]"    "[6.68,9.9]"   
#> [37] "[6.68,9.9]"    "[-1.29,0.814)" "[6.68,9.9]"    "[6.68,9.9]"   
bin_by_ntile(nbins=4)(x)
#>  [1] 1 1 1 2 2 2 2 1 1 2 2 3 2 1 1 3 3 2 3 1 3 4 3 4 2 3 1 3 3 4 4 4 2 3 4 4 4 1
#> [39] 4 4

# \donttest{
bin_by_pam(nbins=4)(x)
#>  [1] 0.4366996 0.4366996 0.4366996 0.4366996 0.4366996 0.4366996 0.4366996
#>  [8] 0.4366996 0.4366996 0.4366996 0.4366996 4.1682095 4.1682095 0.4366996
#> [15] 0.4366996 6.6749515 6.6749515 4.1682095 4.1682095 0.4366996 4.1682095
#> [22] 8.1942194 4.1682095 6.6749515 4.1682095 6.6749515 0.4366996 4.1682095
#> [29] 6.6749515 8.1942194 8.1942194 8.1942194 4.1682095 4.1682095 6.6749515
#> [36] 8.1942194 8.1942194 0.4366996 8.1942194 8.1942194
bin_by_classInt("pretty", nbins=4)(x)
#>  [1] "[0,2)"  "[0,2)"  "[-2,0)" "[2,4)"  "[0,2)"  "[0,2)"  "[0,2)"  "[-2,0)"
#>  [9] "[0,2)"  "[0,2)"  "[0,2)"  "[4,6)"  "[2,4)"  "[0,2)"  "[0,2)"  "[6,8)" 
#> [17] "[4,6)"  "[2,4)"  "[4,6)"  "[-2,0)" "[4,6)"  "[8,10]" "[4,6)"  "[6,8)" 
#> [25] "[2,4)"  "[6,8)"  "[-2,0)" "[4,6)"  "[6,8)"  "[8,10]" "[6,8)"  "[6,8)" 
#> [33] "[2,4)"  "[4,6)"  "[6,8)"  "[8,10]" "[6,8)"  "[-2,0)" "[8,10]" "[8,10]"
# }