Plot a tidyvpcobj

Use ggplot2 graphics to plot and customize the appearance of VPC.

# S3 method for tidyvpcobj
plot(
  x,
  ...,
  facet = FALSE,
  show.points = TRUE,
  show.boundaries = TRUE,
  show.stats = !is.null(x$stats),
  show.binning = isFALSE(show.stats),
  xlab = NULL,
  ylab = NULL,
  color = c("red", "blue", "red"),
  linetype = c("dotted", "solid", "dashed"),
  legend.position = "top",
  facet.scales = "free",
  custom.theme = "ggplot2::theme_bw"
)

Arguments

x
A tidyvpcobj.

...
Further arguments can be specified but are ignored.

facet
Should the resulting plot automatically facet by category? Only applicable for categorical VPC.

show.points
Should the observed data points be plotted?

show.boundaries
Should the bin boundary be displayed?

show.stats
Should the VPC stats be displayed?

show.binning
Should the binning be displayed by coloring the observed data points by bin?

xlab
A character label for the x-axis.

ylab
A character label for the y-axis.

color
A character vector of colors for the percentiles, from low to high.

linetype
A character vector of line type for the percentiles, from low to high.

legend.position
A character string specifying the position of the legend.

facet.scales
A character string specifying the scales argument to use for faceting.

custom.theme
A character string specifying theme from ggplot2 package.

Value

A ggplot object.

See also